Refactor steps into jobs. Add healthcheck for DB. (#498)
Secrets need to be in an environment bound to protected branches to
maintain confidentiality, so we need an independent job for deploy so
that the other CI stages can run during pull requests.

To ensure the same image that is built and tested is the one ultimately
deployed, this uses image save/load and artifact upload/download.

Adds a healthcheck for the database so that the server container doesn't
start until the database is healthy. This should eliminate test
flakiness detected during refactoring.